+++ x/pylons/client/cli/atomgachi_test.go
@@ -69,8 +69,7 @@ func createAtomgachiCookbook(t *testing.T, simInfo *atomgachiBasicSim) {
 		"Cookbook for testing demo atomgachi transactions",
 		"Pylons Inc",
 		"v0.0.1",
-		"alex@shmeeload.xyz",
-		"{\"denom\": \"pylons\", \"amount\": \"12\"}",
+		"alex@test.xyz",
 		"true",
 	}
 
@@ -123,6 +122,7 @@ func createMintRecipe(t *testing.T, simInfo *atomgachiBasicSim) {
 		string(entries),
 		string(itemOutputs),
 		"0",
+		"{\"denom\": \"pylons\", \"amount\": \"12\"}",
 		"true",
 		"extraInfo",
 	}