+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 # Changelog
 
-## Unreleased
 
+## Unreleased 
+ 
 ### State Machine Breaking Changes:
 
+- [#517](https://github.com/Pylons-tech/pylons/pull/517) Move the `CostPerBlock` field from Cookbook messages to Recipe messages.
 - [#508](https://github.com/Pylons-tech/pylons/pull/508) Remove the `MinNameFieldLength` and `MinDescriptionFieldLength` params from the chain.
 
 ### Fixes:
 
+- [#519](https://github.com/Pylons-tech/pylons/pull/519) Remove potentially panic-ing array access in `x/pylons/keeper/complete_pending_execution.go`.
 - [#509](https://github.com/Pylons-tech/pylons/pull/509) Bump `github.com/opencontainers/image-spec` to v1.0.2 to avoid security issue.
 
 ### Changes: